seshat: Move the seshat config types into a separate file.
